+++ changelog.md
@@ -7,6 +7,7 @@
 - Updated dependencies.
 - More reliable `useGraphQL` React hook `loadOnMount` option implementation that fixes ESLint `react-hooks/exhaustive-deps` rule errors.
 - Use `function` instead of `const` declarations in places to simplify transpiled output.
+- `GraphQL.reset()` test name typo fix.
 
 ## 8.0.2